migrate_kv_access_policies/requirements.txt
azure-identity
azure-mgmt-keyvault
azure-mgmt-authorization

pyyaml